app/requirements.txt
alembic==1.12.1
bcrypt==4.0.1
blinker==1.7.0
click==8.1.7
colorama==0.4.6
Flask==3.0.0
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.3.0
MarkupSafe==2.1.3
pyotp==2.9.0
pypng==0.20220715.0
python-decouple==3.8
qrcode==7.4.2
SQLAlchemy==2.0.23
typing_extensions==4.8.0
Werkzeug==3.0.1
WTForms==3.1.1
flask-ckeditor==0.5.1
gunicorn==20.1.0
cryptography==42.0.2
bleach==6.1.0